# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: 3GPP Server API
  description: |
    A JSON-to-NGAP/NAS translation server for conformance testing, fuzzing,
    and security assessment of 5G core networks.

    The server manages gNB and UE contexts, translates JSON requests into
    NGAP/NAS messages over SCTP, and returns decoded responses as JSON.

    ## Workflow

    1. **Create a gNB** — dials SCTP to the AMF, performs NG Setup.
    2. **Create a UE** — provisions credentials (K, OPc), computes SUCI.
    3. **Send NGAP messages** — builds NAS inside NGAP, sends over SCTP,
       returns the decoded AMF response.

    ## IE Control

    The JSON schema mirrors NGAP/NAS IEs 1:1. Every optional IE can be
    individually set, overridden, or omitted — enabling both correct
    procedures and deliberate deviations for fuzz testing.

    ## 3GPP References

    - NGAP: TS 38.413
    - NAS 5GS: TS 24.501
    - 5G Security: TS 33.501
  version: "0.2.0"
  license:
    name: BUSL-1.1
    identifier: BUSL-1.1
paths:
...
  /gnb:
    post:
      operationId: createGnB
      summary: Create gNB
      description: |
        Dials SCTP to the AMF, sends an NGSetupRequest, and returns the
        NGSetupResponse or NGSetupFailure.

        Use `ng_setup_ies` to send a hand-crafted IE list instead of
        auto-building from the top-level fields. Set `skip_ng_setup` to open the
        association without NG Setup (no NGSetupResponse is returned).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGnBRequest"
      responses:
        "201":
          description: gNB created and NG Setup completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGnBResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /gnb/{gnb_id}:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
    get:
      operationId: getGnB
      summary: Get gNB state
      responses:
        "200":
          description: gNB state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GnBStateResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGnB
      summary: Delete gNB
      description: Closes the SCTP connection and removes the gNB.
      responses:
        "204":
          description: gNB deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /gnb/{gnb_id}/gtpu/echo:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
    post:
      operationId: sendGTPUEcho
      summary: Send a GTP-U Echo Request to the UPF (N3 path management)
      description: |
        Sends a GTP-U Echo Request (TS 29.281 §7.2.1) from the gNB's N3 endpoint
        to upf_ip and returns the Echo Response. Requires the gNB to have been
        created with enable_gtpu.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upf_ip]
              properties:
                upf_ip: { type: string, description: "UPF N3 IP to echo." }
                timeout_ms: { type: integer, description: "How long to wait. Default 5000." }
      responses:
        "200": { description: "Echo Response received." }
        "504": { description: "No Echo Response before the timeout." }
  /gnb/{gnb_id}/gtpu/error-indication/await:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
    post:
      operationId: awaitErrorIndication
      summary: Wait for a GTP-U Error Indication from the UPF
      description: |
        Blocks for a GTP-U Error Indication, which the UPF must send when it
        receives a G-PDU for a TEID with no context (TS 29.281 §7.3.1). Requires
        the gNB to have been created with enable_gtpu.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                timeout_ms: { type: integer, description: "How long to wait. Default 5000." }
      responses:
        "200": { description: "Error Indication received." }
        "504": { description: "No Error Indication before the timeout." }
  /gnb/{gnb_id}/ue/{ue_id}/tunnel:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    get:
      operationId: getTunnel
      summary: Get the N3 GTP-U tunnel state for a UE's PDU session
      description: |
        Returns the captured tunnel: the gNB downlink endpoint (n3_gnb_ip,
        dl_teid), the UPF uplink tunnel (upf_ip, ul_teid), the UE IP (ue_ip), and
        the QFI. Populated when the PDU session was established on a GTP-U gNB.
      responses:
        "200": { description: "Tunnel state." }
        "400": { description: "No established N3 tunnel / gNB has no GTP-U endpoint." }
  /gnb/{gnb_id}/ue/{ue_id}/uplink:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: sendUplink
      summary: Send a user-plane uplink packet through the N3 tunnel
      description: |
        Builds an inner IP packet sourced from the UE IP and sends it to the UPF
        as a GTP-U G-PDU (with the uplink PDU Session Container / QFI, TS 38.415).
        Exactly one of icmp_echo or udp is used.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pdu_session_id: { type: integer, description: "Defaults to the UE's session." }
                teid: { type: integer, description: "Override the uplink TEID, e.g. to send to a TEID with no PDR (TS 29.281 §7.3.1)." }
                src: { type: string, description: "Override the inner source IP (source-spoofing tests); defaults to the session's UE IP." }
                icmp_echo:
                  type: object
                  properties:
                    dst: { type: string }
                    id: { type: integer }
                    seq: { type: integer }
                udp:
                  type: object
                  properties:
                    dst: { type: string }
                    dst_port: { type: integer }
                    src_port: { type: integer }
                    payload_hex: { type: string }
      responses:
        "200": { description: "Packet encapsulated and sent." }
        "400": { description: "No established N3 tunnel, or invalid request." }
  /gnb/{gnb_id}/ue/{ue_id}/downlink/await:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: awaitDownlink
      summary: Await a downlink user-plane packet on the N3 tunnel
      description: |
        Blocks for a downlink G-PDU on the session's DL TEID and returns the
        decapsulated inner IP packet (parsed fields under `inner`, plus raw_hex).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pdu_session_id: { type: integer }
                timeout_ms: { type: integer, description: "How long to wait. Default 5000." }
      responses:
        "200": { description: "Decapsulated downlink packet." }
        "504": { description: "No downlink before the timeout." }
...
components:
  parameters:
    gnb_id:
      name: gnb_id
      in: path
      required: true
      schema:
        type: string
    ue_id:
      name: ue_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: SCTP or NGAP error communicating with the AMF.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    CreateGnBRequest:
      type: object
      required: [amf_address]
      properties:
        amf_address:
          type: string
          description: "AMF SCTP address (host:port)."
          example: "10.3.0.2:38412"
        gnb_n2_address:
          type: string
          description: Local bind address for SCTP. Defaults to 0.0.0.0.
        mcc:
          type: string
          pattern: "^\\d{3}$"
          example: "001"
        mnc:
          type: string
          pattern: "^\\d{2,3}$"
          example: "01"
        tac:
          type: string
          description: Tracking Area Code (hex).
          example: "000001"
        gnb_id:
          type: string
          description: gNB ID (hex).
          example: "000001"
        name:
          type: string
        sst:
          type: integer
          description: Slice/Service Type.
        sd:
          type: string
          description: Slice Differentiator (hex, optional).
        slices:
          type: array
          items:
            $ref: "#/components/schemas/SliceInput"
        ng_setup_ies:
          type: array
          description: Override — send these raw IEs instead of auto-building.
          items:
            type: object
        skip_ng_setup:
          type: boolean
          description: >
            Open the SCTP association without sending an NG Setup Request,
            modelling an NG-RAN node that has not completed NG Setup (TS 38.413
            §8.7.1.1). The response carries no ng_setup_response.
        raw_ngap_pdu:
          type: string
          description: Hex NGAP PDU sent verbatim instead of an NG Setup Request.
        wait_for:
          type: array
          description: NGAP message types to wait for. Default NGSetupResponse/NGSetupFailure/ErrorIndication.
          items: { type: string }
        timeout_ms:
          type: integer
          description: How long to wait for the response. Default 5000.
        enable_gtpu:
          type: boolean
          description: >
            Bind an N3 GTP-U endpoint so the gNB terminates the user-plane data
            path (send uplink, receive downlink). Only one GTP-U-enabled gNB can
            exist per N3 IP (it binds the GTP-U port).
        gnb_n3_address:
          type: string
          description: >
            N3 IP the GTP-U endpoint binds and advertises as its downlink
            endpoint. Defaults to gnb_n2_address.
    SliceInput:
      type: object
      properties:
        sst:
          type: integer
        sd:
          type: string
    CreateGnBResponse:
      type: object
      properties:
        gnb_id:
          type: string
        ng_setup_response:
          $ref: "#/components/schemas/NGAPResponse"
    GnBStateResponse:
      type: object
      properties:
        id:
          type: string
        mcc:
          type: string
        mnc:
          type: string
        tac:
          type: string
        gnb_id:
          type: string
        name:
          type: string
        sst:
          type: integer
        sd:
          type: string
...
    NGAPResponse:
      type: object
      properties:
        pdu_type:
          type: string
          enum: [initiating_message, successful_outcome, unsuccessful_outcome]
        message_type:
          type: string
          description: "Decoded message name (e.g. NGSetupResponse, DownlinkNASTransport)."
        raw_hex:
          type: string
          description: Full NGAP PDU as hex.
        ies:
          type: array
          items:
            type: object
            description: Decoded IE with typed fields.
...